requirements.txt
fastapi>=0.121.0
pydantic-ai>=1.12.0
pydantic-ai-slim[google,openai]>=1.12.0
python-dotenv>=1.0.0
uvicorn>=0.38.0
